+++ BankAccountTransactions.Data/Migrations/BankAccountTransactionsContextModelSnapshot.cs
@@ -35,7 +35,7 @@ namespace BankAccountTransactions.Data.Migrations
                         .HasColumnType("varchar(18)")
                         .HasColumnName("accountNumber");
 
-                    b.Property<double>("Balance")
+                    b.Property<decimal>("Balance")
                         .HasColumnType("numeric")
                         .HasColumnName("balance");
 
@@ -61,7 +61,7 @@ namespace BankAccountTransactions.Data.Migrations
                         .HasColumnName("id")
                         .HasDefaultValueSql("gen_random_uuid()");
 
-                    b.Property<double>("Amount")
+                    b.Property<decimal>("Amount")
                         .HasColumnType("numeric")
                         .HasColumnName("amount");
